Extracted getToken to own util file.
