+++ src/Components/Dropbox.tsx
@@ -1,8 +1,9 @@
 import React, { useState } from "react";
 import CircularProgress from "@mui/material/CircularProgress";
 import { useLanguage } from "./LanguageContext";
-import { blobHandler } from "./blobCreator";
+import { blobHandler } from "./BlobCreator";
 import { useMsal } from "@azure/msal-react";
+import { getToken } from "./GetToken";
 
 interface DropboxProps {
   name: string;
@@ -14,22 +15,6 @@ const Dropbox: React.FC<DropboxProps> = ({ name, fetchEndpoint }) => {
   const [isLoading, setIsLoading] = useState(false);
   const { instance, accounts } = useMsal();
   const account = accounts[0];
-
-  const getToken = async () => {
-    const clientId = import.meta.env.VITE_API_AZURE_CLIENT_ID;
-    const defaultScope = `api://${clientId}/user_impersonation`;
-    const request = {
-      scopes: [defaultScope, "User.Read"],
-      account: account,
-    };
-    try {
-      const response = await instance.acquireTokenSilent(request);
-      return response.accessToken;
-    } catch (error) {
-      const popupResponse = await instance.acquireTokenPopup(request);
-      return popupResponse.accessToken;
-    }
-  };
 
   const [error, setError] = useState<string | null>(null);
 
@@ -55,7 +40,7 @@ const Dropbox: React.FC<DropboxProps> = ({ name, fetchEndpoint }) => {
         formData.append("file", data);
         setIsLoading(true);
         setError(null);
-        const token = await getToken();
+        const token = await getToken(instance, account);
         const response = await fetch(fetchEndpoint, {
           method: "POST",
           body: formData,
@@ -91,7 +76,7 @@ const Dropbox: React.FC<DropboxProps> = ({ name, fetchEndpoint }) => {
         formData.append("file", data);
         setIsLoading(true);
         setError(null);
-        const token = await getToken();
+        const token = await getToken(instance, account);
         const response = await fetch(fetchEndpoint, {
           method: "POST",
           body: formData,
