+++ src/Pages/reportspage.tsx
@@ -1,11 +1,12 @@
 import React, { useState } from "react";
 import { CircularProgress } from "@mui/material";
 import Dropbox from "../Components/Dropbox";
-import { blobHandler } from "../Components/blobCreator";
+import { blobHandler } from "../Components/BlobCreator";
 import { useLanguage } from "../Components/LanguageContext";
 import UseButton from "../Components/UseButton";
 import { LastUpdatedText } from "../Components/LastUpdated";
 import { useMsal } from "@azure/msal-react";
+import { getToken } from "../Components/GetToken";
 
 const Reports: React.FC = () => {
   // Fetch template based on dropbox
@@ -17,21 +18,6 @@ const Reports: React.FC = () => {
 
   const { instance, accounts } = useMsal();
   const account = accounts[0];
-  const getToken = async () => {
-    const clientId = import.meta.env.VITE_API_AZURE_CLIENT_ID;
-    const defaultScope = `api://${clientId}/user_impersonation`;
-    const request = {
-      scopes: [defaultScope, "User.Read"],
-      account: account,
-    };
-    try {
-      const response = await instance.acquireTokenSilent(request);
-      return response.accessToken;
-    } catch (error) {
-      const popupResponse = await instance.acquireTokenPopup(request);
-      return popupResponse.accessToken;
-    }
-  };
 
   const handleFetchTemplate = async (button: string) => {
     setLoading((prevState) => ({ ...prevState, [button]: true }));
@@ -58,7 +44,7 @@ const Reports: React.FC = () => {
     setLoading((prevState) => ({ ...prevState, [button]: true }));
     try {
       setError(null);
-      const token = await getToken();
+      const token = await getToken(instance, account);
       const response = await fetch(import.meta.env.VITE_API_EXCELFULLVIEW_URL, {
         headers: {
           Authorization: `Bearer ${token}`,
